fix: send RGB params for WS solid effect

+++ UltraNodeV5/components/ul_ws_engine/effects_ws/solid.c
@@ -6,10 +6,12 @@ void solid_init(void) { (void)0; }
 
 void solid_apply_params(int strip, const cJSON* params) {
     if (!params || !cJSON_IsArray(params) || cJSON_GetArraySize(params) < 3) return;
-    int r = cJSON_GetArrayItem(params, 0)->valueint;
-    int g = cJSON_GetArrayItem(params, 1)->valueint;
-    int b = cJSON_GetArrayItem(params, 2)->valueint;
-    ul_ws_set_solid_rgb(strip, (uint8_t)r, (uint8_t)g, (uint8_t)b);
+
+    uint8_t r = (uint8_t)cJSON_GetArrayItem(params, 0)->valueint;
+    uint8_t g = (uint8_t)cJSON_GetArrayItem(params, 1)->valueint;
+    uint8_t b = (uint8_t)cJSON_GetArrayItem(params, 2)->valueint;
+
+    ul_ws_set_solid_rgb(strip, r, g, b);
 }
 
 void solid_render(uint8_t* frame_rgb, int pixels, int frame_idx) {
